fix: harden data safety, matching correctness, security, and packaging
Codex audited the app and this lands the fix pass, independently
re-verified end-to-end (test suite, lint/type gate, a rebuilt wheel,
and a live GUI smoke drive) before committing.

- Atomic writes for project/settings/CSV/XLSX/PDF outputs; dirty-state
  tracking with save prompts on New/Open/Open Recent/Exit
- Fix legacy vibration-project migration crash, background-match race
  condition, exclusion-band scoring, interpolation zero-fill
  fingerprint artifact, baseline-corrected peak detection prominence
- Batch workflows share peak-detection logic with the interactive
  workflow, gain a background-task close guard, and get per-item DB
  transactions for reference import
- CSV/XLSX formula-injection and PDF markup-injection protections;
  NIST host/redirect/size validation; JSME asset pinning and hashing;
  SQL LIKE escaping for folder-scope queries
- Fix wheel packaging (missing main.py/assets), OS file-association
  handling, version drift across installers, release CI test/lint
  gating, and tag-injection in the release workflow
- JCAMP reader now rejects unsupported/ambiguous encodings instead of
  silently misparsing them
- Remove confirmed-dead code paths and add corresponding regression
  tests

+++ processing/baseline.py
@@ -81,8 +81,9 @@ def _convex_hull_baseline(
 
     # Monotone-chain lower convex hull.
     def _cross(o: int, a: int, b: int) -> float:
-        return (wn_sorted[a] - wn_sorted[o]) * (y_work[b] - y_work[o]) - (y_work[a] - y_work[o]) * (
-            wn_sorted[b] - wn_sorted[o]
+        return float(
+            (wn_sorted[a] - wn_sorted[o]) * (y_work[b] - y_work[o])
+            - (y_work[a] - y_work[o]) * (wn_sorted[b] - wn_sorted[o])
         )
 
     hull: list[int] = []
